requirements.txt
numpy==1.26.4
pandas==2.2.2
PyYAML==6.0.1
Flask==3.0.3
bayesian-optimization==1.4.3
matplotlib==3.8.4
scikit-learn==1.3.2
xgboost==2.0.1
imblearn
seaborn==0.13.2

